+++ Changelog.md
@@ -50,4 +50,5 @@
 - MAJOR code cleanup.
 - Added catch for invalid file name when creating a file.
 - Updated ADR to not retry after ADR completed & dependency couldn't be resolved.
-- Rewrote plugin remove command for better readability & maintainability.
+- Rewrote plugin remove commands for better readability & maintainability.
+- Plugins will be enabled immediately on installation.